Add checkboxes and radio buttons
Radio buttons work a little differently to MacOS in as much as they post notifications regarding that a button in a given radioGroup was checked, and other buttons in that group listen for the notifications and update their state. This seems pretty flexible and a bit better than the NSMatrix approach that Cocoa takes.