Fix a bug where textures could be reassigned wrongly
We were using the max count of the textures array in the renderer as the next texture id, but as well as increasing, this value fluctuates downwards as we remove textures from the array when they're no longer required. That can mean out-of-date keys in the map, and problems ensuing. Making the texture-id always increase solves the problem, and we have a 64-bit texture id, so I don't think we'll run out during the lifetime of the program.