Change when the backing texture is installed
We were installing the backing texture when the view was added to the superview. but that is too early when creating the initial bootstrap view-set from a ZIB because some views won't be attached to their superview yet if they're just custom views in the ZIB top-level. We now do it when the painter is about to paint the view, unless it's already been done.