Fix the tiled-render problem for AZSegmentedControl
The issue here was that the tiling only has control of what to do in both directions simultaneously. We want stretch-in-Y and repeat-in-X and there is no way to do that with SDL_RenderTextureTiled. Instead what we do is create a frame-sized texture for all the states (6 of them), create a temporary single texture composite of all the UI parts together, and use a 9-way blit to fill out the frame-sized texture. Then we can blit from the frame-sized texture simply in the draw routine. This gets rid of the ugly lower border and makes it look nice and neat.