Patch for delegate to implement scrolling when there's a selection
This is beginning to feel over-complicated for something as simple as selection. The complexity comes from recycling the subviews that make up the table, but OutlineView gets around it by simply creating new views every time it needs to relayout. Perhaps TableView ought to do the same, then all the selection logic is in the 'createView' method called on the delegate. Right now it's spread out over 3 methods

- the selection is *about* to change, so unselect anything currently selected
- the selection *did* change, so reselect anything now selected
- we're about to draw a view, which might previously have been selected, so make sure its selected state is still valid,

This is partly due to the view-based tableView as well - the tableView itself doesn't know how to "select" something - it could be turn-red, it could be invert-the-colours, it could be anything. So it has to tell the delegate sufficient info that the delegate (which *does* know what to do) can do it..