Alter the delegate protocol
It seems the outline view transform is invalid if the click is outside the bounds of the outline view but still inside the bounds of the enclosing scroll/clip views on-screen. Cope with this by using the superview transform in this case. This necessitates sending the *actual* view used as well as the outlineview object, even if they are the same...