First pass at multiple window management
Looks like it's fundamentially broken because SDL's Render API associates a renderer with a window, to manage the swapchain. That means the renderer is per-window, which means you need a window to get access to the renderer.

The problem is that we often have stuff in -initWithFrame: which would call on the renderer, like "create font" or "create image". And that view isn't in the view hierarchy yet (because it's still being created) so we don't have the window context of the parent. That association is made in -addSubview

We *could* make the default view initialiser -initWithFrame:inWindow: which would solve the problem, but that is ... a lot of replacement work... for something that's not exactly high on the priority list...